EditVideoAssignment component done

+++ lib/utils.ts
@@ -46,7 +46,7 @@ export const formatDateTime = (dateString: Date) => {
   };
 
   const dateOptions: Intl.DateTimeFormatOptions = {
-    month: "long", // abbreviated month name (e.g., 'Oct')
+    month: "short", // abbreviated month name (e.g., 'Oct')
     year: "numeric", // numeric year (e.g., '2023')
     day: "numeric", // numeric day of the month (e.g., '25')
   };
